working checkout, fixed some bugs, added bike seeder, added rating table, updated order table

+++ app/Http/Controllers/BasketController.php
@@ -2,10 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\DTO\Basket;
-use App\DTO\BasketItem;
 use App\Http\Controllers\Controller;
-use App\Models\Auction;
 use App\Models\Bike;
 use App\Models\Order;
 use App\Models\OrderItem;
@@ -13,7 +10,6 @@ use App\Services\Basket\BasketManager;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Redirect;
-use Illuminate\Support\Facades\Session;
 use Inertia\Inertia;
 
 class BasketController extends Controller
@@ -34,12 +30,6 @@ class BasketController extends Controller
         } catch (\Exception $e) {
             return Redirect::route('rent', ['id' => $bike->id])->with('error', $e->getMessage());
         }
-        
-        // $this->basketManager->changeQuantity(2, 1);
-
-        // var_dump($this->basketManager->getBasket());
-        // var_dump($this->basketManager->getBasket()->getTotalPrice());
-
 
         return Redirect::route('checkout')->with('success', 'Item added successfully');
     }
@@ -47,7 +37,6 @@ class BasketController extends Controller
     public function getBasket()
     {
         $basket = $this->basketManager->getBasket();
-        // dd($basket);
 
         return Inertia::render('Checkout',[
                 'basket' => $basket,
@@ -55,40 +44,62 @@ class BasketController extends Controller
         );
     }
 
+    public function removeItem($id)
+    {
+        $basket = $this->basketManager->getBasket();
 
+        if($basket)
+        {
+            $this->basketManager->removeItem($id);
+            return Redirect::route('checkout')->with('success', 'Item deleted successfully');
+        }
+        return Redirect::route('checkout')->with('error', 'Item deleted unsuccessfully');
+    }
 
     public function store(Request $request)
     {   
-        
         $basket = $this->basketManager->getBasket();
 
-        $user = $request->user();
-
-        $orderNumberid = DB::table('order_sequence')->insertGetId(['id' => null]);
-        DB::table('order_sequence')->where('id', '<', $orderNumberid)->delete();
-
-
-        $order = Order::create([
-            'user_id' => $user->id,
-            'total_price' => $basket->getTotalPrice(),
-        ]);
-
-        $order->order_nr = '#' . str_pad($orderNumberid, 4, "0", STR_PAD_LEFT);
-        $order->save();
-
-        foreach ($basket->getItems() as $item) {
-            OrderItem::create([
-                'bike_id' => $item->getProductId(),
-                'order_id' => $order->id,
-                'quantity' => $item->getQuantity(), // pravaliduoti pries visus save quantity ar turim in_stock
-                'unit_price' => $item->getUnitPrice(),
-            ]);
+        if($basket)
+        {
+            foreach ($basket->getItems() as $item) {
+                $user = $request->user();
+        
+                $orderNumberid = DB::table('order_sequence')->insertGetId(['id' => null]);
+                DB::table('order_sequence')->where('id', '<', $orderNumberid)->delete();
+        
+                $bike = Bike::find($item->getProductId());
+                
+                if($bike->in_stock >= $item->getQuantity())
+                {
+                    $order = Order::create([
+                        'user_id' => $user->id,
+                        'total_price' => $basket->getTotalPrice(),
+                    ]);
+        
+                    $order->order_nr = '#' . str_pad($orderNumberid, 4, "0", STR_PAD_LEFT);
+                    
+                    OrderItem::create([
+                        'bike_id' => $item->getProductId(),
+                        'order_id' => $order->id,
+                        'quantity' => $item->getQuantity(),
+                        'unit_price' => $item->getUnitPrice(),
+                    ]);
+        
+                    $bike->update(['in_stock' => $bike->in_stock - $item->getQuantity()]);
+        
+                    $order->save();
+        
+                    $this->basketManager->clear();
+                    
+                    return Redirect::route('dashboard')->with('success', 'Order successfully confirmed.');
+                }else{
+                    $this->basketManager->clear();
+        
+                    return Redirect::route('dashboard')->with('error', 'Order cannot be confirmed, because not enough quantity. Try again later.');
+                }
+            } 
         }
-
-        //pabaigoje sumazinti in_stock kieki.
-
-        $this->basketManager->clear();
-
-        return Redirect::route('dashboard')->with('success', 'User created successfully');
+        return Redirect::route('dashboard')->with('error', 'Order cannot be confirmed, because no items in basket.');
     }
 }
